MindSpaceApi/requirements.txt
Django>=3.2,<3.3
djangorestframework>=3.12.4,<3.13.0
mysqlclient>=2.0.3,<2.1.0
firebase-admin>=4.5.3,<4.6.0
django-cors-headers>=3.7.0,<3.8.0
nltk>=3.6.2,<3.7.0